+++ changelog.md
@@ -1,6 +1,7 @@
 ### 4.2.0
 - Added Mixin blaming
 - Added Chinese translation
+- Fixed crashes not being caught in Forge in certain cases (1.18 bug) 
 ### 4.1.6
 - Updated to support Quilt hashed mappings.
 ### 4.1.5